requirements.txt
# FastAPI và Web Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Ngrok
pyngrok==7.0.1

# Async utilities
nest-asyncio==1.5.8
aiohttp==3.9.1

# Data validation
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0